+++ src/main/scala/io/iohk/ethereum/network/p2p/PV63.scala
@@ -1,0 +1,5 @@
+package io.iohk.ethereum.network.p2p
+
+package object PV63 {
+
+}
